+++ Setup-Your-Mac-via-Dialog.bash
@@ -73,8 +73,9 @@
 #   Version 1.12.10, 15-Sep-2023, Dan K. Snelson (@dan-snelson)
 #   - Better WelcomeMessage logic and variable handling ([Pull Request No. 101](https://github.com/dan-snelson/Setup-Your-Mac/pull/101); thanks big bunches, @GadgetGeekNI!)
 #
-#   Version 1.12.11, 25-Sep-2023, Dan K. Snelson (@dan-snelson)
+#   Version 1.12.11, 26-Sep-2023, Dan K. Snelson (@dan-snelson)
 #   - Restored logging of `jamfProPolicyNameFailures`
+#   - Updated `if … then` statements when disabling the "Continue" button in the User Input "Welcome" dialog until Dynamic Download Estimates have complete ([Pull Request No. 115](https://github.com/dan-snelson/Setup-Your-Mac/pull/116); thanks, @delize!)
 #
 ####################################################################################################